WORK IN PROGRESS. renamed interface to helper utils, fixed incorrect types. Working on stamina estimation helper method, cleaned up the intializeFANN helper clusterfuck, more additions to memory, dont attack when right in enemy face, added weapon data to store info about weapons.

+++ Dark Souls AI C/CharacterStruct.h
@@ -21,7 +21,7 @@ typedef struct {
 	float rotation;
 	//current animation type id
 	ullong animationType_address;
-	unsigned int animationType_id;
+    unsigned short animationType_id;
 	//hp
 	ullong hp_address;
 	unsigned int hp;
